Fix undefined lt_args_wrap in libltaudit when using flex >= 2.6.3
Applies a solution similar to the used in 81e78d8c in the src/args-flex.l
file. Although this undefined symbol does not prevent the build, the
lack of it in the generated libltaudit.so will cause errors such as
"ld.so: object '/latrace-plus/libltaudit.so.0.5.13' cannot be loaded as
audit interface: undefined symbol: le_args_wrap; ignored." when latrace
is invoked with the -A switch.